+++ src/main/java/com/mclegoman/perspective/mixin/client/textured_entity/minecraft/frog/FrogEntityRendererMixin.java
@@ -7,10 +7,13 @@
 
 package com.mclegoman.perspective.mixin.client.textured_entity.minecraft.frog;
 
+import com.google.gson.JsonObject;
 import com.mclegoman.perspective.client.textured_entity.TexturedEntity;
+import com.mclegoman.perspective.common.util.IdentifierHelper;
 import net.minecraft.entity.Entity;
 import net.minecraft.entity.passive.FrogEntity;
 import net.minecraft.util.Identifier;
+import net.minecraft.util.JsonHelper;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
@@ -18,9 +21,24 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 @Mixin(priority = 100, value = net.minecraft.client.render.entity.FrogEntityRenderer.class)
 public class FrogEntityRendererMixin {
-	@Inject(at = @At("RETURN"), method = "getTexture(Lnet/minecraft/entity/Entity;)Lnet/minecraft/util/Identifier;", cancellable = true)
-	private void perspective$getTexture(Entity entity, CallbackInfoReturnable<Identifier> cir) {
-		if (entity instanceof FrogEntity)
-			cir.setReturnValue(TexturedEntity.getTexture(entity, "minecraft:frog", cir.getReturnValue()));
+	@Inject(at = @At("RETURN"), method = "getTexture(Lnet/minecraft/entity/passive/FrogEntity;)Lnet/minecraft/util/Identifier;", cancellable = true)
+	private void perspective$getTexture(FrogEntity entity, CallbackInfoReturnable<Identifier> cir) {
+		if (entity != null) {
+			boolean isTexturedEntity = true;
+			JsonObject entitySpecific = TexturedEntity.getEntitySpecific(entity, "minecraft:frog");
+			if (entitySpecific != null) {
+				String type = entity.getVariant().getIdAsString().toLowerCase();
+				if (entitySpecific.has(type)) {
+					JsonObject typeRegistry = JsonHelper.getObject(entitySpecific, entity.getVariant().getIdAsString().toLowerCase());
+					if (typeRegistry != null) isTexturedEntity = JsonHelper.getBoolean(typeRegistry, "enabled", true);
+				}
+			}
+			if (isTexturedEntity) {
+				String variant = entity.getVariant().getIdAsString();
+				String variantNamespace = IdentifierHelper.getStringPart(IdentifierHelper.Type.NAMESPACE, variant);
+				String variantKey = IdentifierHelper.getStringPart(IdentifierHelper.Type.KEY, variant);
+				cir.setReturnValue(TexturedEntity.getTexture(entity, variantNamespace, "minecraft:frog", TexturedEntity.Affix.SUFFIX, "_" + variantKey, cir.getReturnValue()));
+			}
+		}
 	}
 }